requirements.txt
aiobotocore==2.5.0
aiohttp==3.8.5
aioitertools==0.11.0
aiosignal==1.3.1
aiosqlite==0.19.0
alembic==1.10.3
anyio==3.6.2
apprise==1.3.0
asgi-lifespan==2.1.0
async-timeout==4.0.2
asyncpg==0.27.0
attrs==23.1.0
botocore==1.29.76
cachetools==5.3.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==2.2.1
colorama==0.4.6
convertdate==2.4.0
coolname==2.2.0
croniter==1.3.14
cryptography==41.0.2
dateparser==1.1.8
docker==6.0.1
duckdb==0.7.1
et-xmlfile==1.1.0
fastapi==0.95.1
frozenlist==1.3.3
fsspec==2023.4.0
google-auth==2.17.3
greenlet==2.0.2
griffe==0.27.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.17.0
httpx==0.24.0
hyperframe==6.0.1
idna==3.4
Jinja2==3.1.2
jmespath==1.0.1
jsonpatch==1.32
jsonpointer==2.3
jsonschema==4.17.3
kubernetes==26.1.0
lunardate==0.2.0
Mako==1.2.4
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
mdurl==0.1.2
minio==7.1.14
multidict==6.0.4
numpy==1.24.2
oauthlib==3.2.2
openpyxl==3.1.2
orjson==3.8.10
packaging==23.1
pandas==2.0.0
pathspec==0.11.1
pendulum==2.1.2
prefect==2.10.4
pyarrow==11.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.7
Pygments==2.15.1
pyluach==2.2.0
PyMeeus==0.5.12
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-slugify==8.0.1
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
pytzdata==2020.1
PyYAML==6.0
readchar==4.0.5
regex==2023.3.23
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.3.4
rsa==4.9
s3fs==2023.4.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.47
starlette==0.27.0
text-unidecode==1.3
toml==0.10.2
typer==0.7.0
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
urllib3==1.26.15
uvicorn==0.21.1
websocket-client==1.5.1
websockets==11.0.2
workalendar==17.0.0
wrapt==1.15.0
yarl==1.9.2
pydriller==2.5.1
